+++ src/main/java/br/com/feirasverdes/backend/service/EstandeService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -15,6 +16,8 @@ import br.com.feirasverdes.backend.dto.EstandeDto;
 import br.com.feirasverdes.backend.entidade.Endereco;
 import br.com.feirasverdes.backend.entidade.Estande;
 import br.com.feirasverdes.backend.entidade.Imagem;
+import br.com.feirasverdes.backend.exception.EstandeNaoPertenceAoUsuarioException;
+import br.com.feirasverdes.backend.exception.FeiraNaoPertenceAoUsuarioException;
 
 @Service
 public class EstandeService {
@@ -25,9 +28,12 @@ public class EstandeService {
 	@Autowired
 	private EnderecoDao enderecoDao;
 
-	public void atualizarEstande(final Long id, final EstandeDto estandeAtualizado) throws IOException {
-		// TODO verificar se usuário chamando método foi quem criou
+	public void atualizarEstande(final Long id, final EstandeDto estandeAtualizado) throws IOException, EstandeNaoPertenceAoUsuarioException {
 		Estande estande = dao.getOne(id);
+		
+		if (!estande.getUsuario().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+			throw new EstandeNaoPertenceAoUsuarioException("Esta feira não foi cadastrada por você.");
+		}
 
 		if (estandeAtualizado.getImagem() != null) {
 			Imagem imagem = new Imagem();
@@ -65,8 +71,12 @@ public class EstandeService {
 		return dao.save(estande);
 	}
 
-	public void excluirEstande(Long id) {
-		// TODO verificar se usuário chamando método foi quem criou
+	public void excluirEstande(Long id) throws EstandeNaoPertenceAoUsuarioException {
+		Estande estande = dao.getOne(id);
+		
+		if (!estande.getUsuario().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+			throw new EstandeNaoPertenceAoUsuarioException("Esta feira não foi cadastrada por você.");
+		}
 		dao.deleteById(id);
 	}
